Curso - Aula - Questao
Estas três classes estão cadastrando buscando e excluindo.

+++ Efeito Joule/src/br/com/joule/dao/QuestaoDAO.java
@@ -9,6 +9,6 @@ public interface QuestaoDAO extends DAO<Questao, Long> {
 
 	Questao buscarPorPergunda(String pergunta);
 
-	List<Questao> buscarPorAula(int codigo);
-
+	List<Questao> buscarPorAula(long codigo);
+	
 }
